Renamed PDFTest, moved file input into FileUpload
